+++ packages/rainbowkit/src/components/ConnectOptions/ConnectDetails.tsx
@@ -2,6 +2,7 @@ import React, { ReactNode, useContext, useEffect } from 'react';
 import { touchableStyles } from '../../css/touchableStyles';
 import { useWindowSize } from '../../hooks/useWindowSize';
 import { BrowserType, getBrowser, isSafari } from '../../utils/browsers';
+import { getGradientRGBAs } from '../../utils/colors';
 import { InstructionStepName } from '../../wallets/Wallet';
 import {
   useWalletConnectors,
@@ -357,13 +358,16 @@ export function ConnectDetail({
 const DownloadOptionsBox = ({
   actionLabel,
   description,
+  iconAccent,
   iconBackground,
   iconUrl,
   isCompact,
   onAction,
   title,
   url,
+  variant,
 }: {
+  iconAccent?: string;
   title: string;
   description: string;
   onAction?: () => void;
@@ -372,14 +376,15 @@ const DownloadOptionsBox = ({
   isCompact: boolean;
   iconUrl: string | (() => Promise<string>);
   iconBackground?: string;
+  variant: 'browser' | 'app';
 }) => {
+  const isBrowserCard = variant === 'browser';
+  const gradientRgbas =
+    !isBrowserCard && iconAccent && getGradientRGBAs(iconAccent);
   return (
     <Box
       alignItems="center"
-      borderColor="generalBorder"
       borderRadius="13"
-      borderStyle="solid"
-      borderWidth="1"
       display="flex"
       justifyContent="center"
       overflow="hidden"
@@ -389,40 +394,114 @@ const DownloadOptionsBox = ({
       width="full"
     >
       <Box
-        background="generalBorder"
-        height="full"
-        position="absolute"
+        borderColor="actionButtonBorder"
+        borderRadius="13"
+        borderStyle="solid"
+        borderWidth="1"
         style={{
-          filter: 'blur(100px)',
-          transform: 'translate3d(0, 0, 0)',
-          zIndex: 0,
+          bottom: '0',
+          left: '0',
+          position: 'absolute',
+          right: '0',
+          top: '0',
+          zIndex: 1,
         }}
-        width="full"
-      >
+      />
+      {isBrowserCard && (
         <Box
-          display="flex"
-          flexDirection="row"
-          justifyContent="space-between"
+          background="downloadTopCardBackground"
+          height="full"
+          position="absolute"
+          style={{
+            zIndex: 0,
+          }}
+          width="full"
+        >
+          <Box
+            display="flex"
+            flexDirection="row"
+            justifyContent="space-between"
+            style={{
+              bottom: '0',
+              filter: 'blur(20px)',
+              left: '0',
+              position: 'absolute',
+              right: '0',
+              top: '0',
+              transform: 'translate3d(0, 0, 0)',
+            }}
+          >
+            <Box
+              style={{
+                filter: 'blur(100px)',
+                marginLeft: -27,
+                marginTop: -20,
+                opacity: 0.6,
+                transform: 'translate3d(0, 0, 0)',
+              }}
+            >
+              <AsyncImage
+                borderRadius="full"
+                height="200"
+                src={iconUrl}
+                width="200"
+              />
+            </Box>
+            <Box
+              style={{
+                filter: 'blur(100px)',
+                marginRight: 0,
+                marginTop: 105,
+                opacity: 0.6,
+                overflow: 'auto',
+                transform: 'translate3d(0, 0, 0)',
+              }}
+            >
+              <AsyncImage
+                borderRadius="full"
+                height="200"
+                src={iconUrl}
+                width="200"
+              />
+            </Box>
+          </Box>
+        </Box>
+      )}
+      {!isBrowserCard && gradientRgbas && (
+        <Box
+          background="downloadBottomCardBackground"
           style={{
             bottom: '0',
-            filter: 'opacity(30%)',
             left: '0',
             position: 'absolute',
             right: '0',
             top: '0',
-            transform: 'translate3d(0, 0, 0)',
           }}
         >
-          <Box style={{ marginLeft: -100, marginTop: -100 }}>
-            <AsyncImage
-              borderRadius="full"
-              height="360"
-              src={iconUrl}
-              width="360"
-            />
-          </Box>
+          <Box
+            position="absolute"
+            style={{
+              background: `radial-gradient(50% 50% at 50% 50%, ${gradientRgbas[0]} 0%, ${gradientRgbas[1]} 25%, rgba(0,0,0,0) 100%)`,
+              height: 564,
+              left: -215,
+              top: -197,
+              transform: 'translate3d(0, 0, 0)',
+              width: 564,
+            }}
+          />
+          <Box
+            position="absolute"
+            style={{
+              background: `radial-gradient(50% 50% at 50% 50%, ${gradientRgbas[2]} 0%, rgba(0, 0, 0, 0) 100%)`,
+              height: 564,
+              left: -1,
+              top: -76,
+              transform: 'translate3d(0, 0, 0)',
+              width: 564,
+            }}
+          />
         </Box>
-      </Box>
+      )}
       <Box
         alignItems="flex-start"
         display="flex"
@@ -519,10 +598,12 @@ export function DownloadOptionsDetail({
           onAction={() => changeWalletStep(WalletStep.InstructionsExtension)}
           title={`${wallet.name} for ${browser}`}
           url={wallet?.downloadUrls?.browserExtension}
+          variant="browser"
         />
         <DownloadOptionsBox
           actionLabel="Get the app"
           description="Use the mobile wallet to explore the world of Ethereum."
+          iconAccent={wallet.iconAccent}
           iconBackground={wallet.iconBackground}
           iconUrl={wallet.iconUrl}
           isCompact={isCompact}
@@ -530,6 +611,7 @@ export function DownloadOptionsDetail({
             changeWalletStep(WalletStep.Download);
           }}
           title={`${wallet.name} for Mobile`}
+          variant="app"
         />
       </Box>
     </Box>