+++ packages/create-rainbowkit/templates/next-app/CHANGELOG.md
@@ -1,5 +1,21 @@
 # rainbowkit-next-app
 
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies [2e6bb8f]
+- Updated dependencies [2e6bb8f]
+  - @rainbow-me/rainbowkit@0.7.0
+
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies [ecd7209]
+- Updated dependencies [248a1cb]
+  - @rainbow-me/rainbowkit@0.6.2
+
 ## 0.0.26
 
 ### Patch Changes