+++ packages/rainbowkit/CHANGELOG.md
@@ -1,5 +1,59 @@
 # @rainbow-me/rainbowkit
 
+## 0.7.0
+
+### Minor Changes
+
+- 2e6bb8f: Support tree shaking of individual wallets by exposing them as separate imports via the new `@rainbow-me/rainbowkit/wallets` entry point.
+
+  In order to reduce bundle size, you can now select the individual wallets you want to import into your application.
+
+  Note that since wallets are no longer namespaced via the `wallet` object, all wallets now have a `Wallet` suffix.
+
+  ```diff
+  -import { connectorsForWallets, wallet } from '@rainbow-me/rainbowkit';
+  +import { connectorsForWallets } from '@rainbow-me/rainbowkit';
+  +import {
+  +  injectedWallet,
+  +  rainbowWallet,
+  +  metaMaskWallet,
+  +  coinbaseWallet,
+  +  walletConnectWallet,
+  +} from '@rainbow-me/rainbowkit/wallets';
+
+  const wallets = [
+  -  wallet.injected({ chains }),
+  -  wallet.rainbow({ chains }),
+  -  wallet.metaMask({ chains }),
+  -  wallet.coinbase({ chains, appName: 'My App' }),
+  -  wallet.walletConnect({ chains }),
+  +  injectedWallet({ chains }),
+  +  rainbowWallet({ chains }),
+  +  metaMaskWallet({ chains }),
+  +  coinbaseWallet({ chains, appName: 'My App' }),
+  +  walletConnectWallet({ chains }),
+  ];
+  ```
+
+- 2e6bb8f: The Steakwallet backwards compatibility layer has been removed. Omni should be used instead, available via the new `@rainbow-me/rainbowkit/wallets` entry point.
+
+  ```diff
+  -import { wallet } from '@rainbow-me/rainbowkit';
+  +import { omniWallet } from '@rainbow-me/rainbowkit/wallets';
+
+  const wallets = [
+  -  wallet.steak({ chains }),
+  +  omniWallet({ chains }),
+  ];
+  ```
+
+## 0.6.2
+
+### Patch Changes
+
+- ecd7209: The new Wallet installation flows add two new colors to RainbowKit's `colors` property in the custom `theme` object: `downloadBottomCardBackground` and `downloadTopCardBackground`. Additionally, we now include the optional `iconAccent` prop in the `Wallet` object, specifically for wallets with both a browser extension and an app, to use as the accent color for the new download installation flows.
+- 248a1cb: Generate a new QR code when WalletConnect request is rejected
+
 ## 0.6.1
 
 ### Patch Changes